devops/messaging_system/requirements.txt
fastapi
uvicorn[standard]
celery
amqp
pika
python-dotenv